third_party/blink/web_tests/platform/generic/http/tests/local/formdata/send-form-data-with-empty-blob-filename-expected.txt

Test that filename passed to FormData.append() takes precedence over default Blob filename 'blob', even if empty.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Sending FormData containing one blob with custom empty filename:
PASS filename is ''
PASS successfullyParsed is true

TEST COMPLETE
